requirements.txt
Mastodon.py==1.1.1
tracery==0.1.1